[sdk][typscript] Export transaction builder apis at high level
Closes: #1350
