+++ ecosystem/typescript/sdk/package.json
@@ -54,5 +54,5 @@
     "tweetnacl": "^1.0.3",
     "yarn": "^1.22.18"
   },
-  "version": "1.0.0"
+  "version": "1.1.0"
 }
