+++ ecosystem/typescript/sdk/src/aptos_client.test.ts
@@ -6,22 +6,7 @@ import { AnyObject } from './util';
 import { FAUCET_URL, NODE_URL } from './util.test';
 import { FaucetClient } from './faucet_client';
 import { AptosAccount } from './aptos_account';
-import {
-  ChainId,
-  RawTransaction,
-  ScriptFunction,
-  StructTag,
-  TransactionPayloadScriptFunction,
-  TypeTagStruct,
-  AccountAddress,
-  MultiEd25519PublicKey,
-  Ed25519PublicKey,
-  MultiEd25519Signature,
-  Ed25519Signature,
-} from './transaction_builder/aptos_types';
-import { bcsSerializeUint64, bcsToBytes } from './transaction_builder/bcs';
-import { AuthenticationKey } from './transaction_builder/aptos_types/authentication_key';
-import { SigningMessage, TransactionBuilderMultiEd25519 } from './transaction_builder';
+import { TxnBuilderTypes, TransactionBuilderMultiEd25519, BCS } from './transaction_builder';
 
 test('gets genesis account', async () => {
   const client = new AptosClient(NODE_URL);
@@ -118,14 +103,14 @@ test(
     accountResource = resources.find((r) => r.type === '0x1::Coin::CoinStore<0x1::TestCoin::TestCoin>');
     expect((accountResource.data as any).coin.value).toBe('0');
 
-    const token = new TypeTagStruct(StructTag.fromString('0x1::TestCoin::TestCoin'));
+    const token = new TxnBuilderTypes.TypeTagStruct(TxnBuilderTypes.StructTag.fromString('0x1::TestCoin::TestCoin'));
 
-    const scriptFunctionPayload = new TransactionPayloadScriptFunction(
-      ScriptFunction.natual(
+    const scriptFunctionPayload = new TxnBuilderTypes.TransactionPayloadScriptFunction(
+      TxnBuilderTypes.ScriptFunction.natual(
         '0x1::Coin',
         'transfer',
         [token],
-        [bcsToBytes(AccountAddress.fromHex(account2.address())), bcsSerializeUint64(717)],
+        [BCS.bcsToBytes(TxnBuilderTypes.AccountAddress.fromHex(account2.address())), BCS.bcsSerializeUint64(717)],
       ),
     );
 
@@ -134,14 +119,14 @@ test(
       client.getChainId(),
     ]);
 
-    const rawTxn = new RawTransaction(
-      AccountAddress.fromHex(account1.address()),
+    const rawTxn = new TxnBuilderTypes.RawTransaction(
+      TxnBuilderTypes.AccountAddress.fromHex(account1.address()),
       BigInt(sequnceNumber),
       scriptFunctionPayload,
       1000n,
       1n,
       BigInt(Math.floor(Date.now() / 1000) + 10),
-      new ChainId(chainId),
+      new TxnBuilderTypes.ChainId(chainId),
     );
 
     const bcsTxn = AptosClient.generateBCSTransaction(account1, rawTxn);
@@ -165,16 +150,16 @@ test(
     const account1 = new AptosAccount();
     const account2 = new AptosAccount();
     const account3 = new AptosAccount();
-    const multiSigPublicKey = new MultiEd25519PublicKey(
+    const multiSigPublicKey = new TxnBuilderTypes.MultiEd25519PublicKey(
       [
-        new Ed25519PublicKey(account1.signingKey.publicKey),
-        new Ed25519PublicKey(account2.signingKey.publicKey),
-        new Ed25519PublicKey(account3.signingKey.publicKey),
+        new TxnBuilderTypes.Ed25519PublicKey(account1.signingKey.publicKey),
+        new TxnBuilderTypes.Ed25519PublicKey(account2.signingKey.publicKey),
+        new TxnBuilderTypes.Ed25519PublicKey(account3.signingKey.publicKey),
       ],
       2,
     );
 
-    const authKey = AuthenticationKey.fromMultiEd25519PublicKey(multiSigPublicKey);
+    const authKey = TxnBuilderTypes.AuthenticationKey.fromMultiEd25519PublicKey(multiSigPublicKey);
 
     const mutisigAccountAddress = authKey.derivedAddress();
     await faucetClient.fundAccount(mutisigAccountAddress, 5000);
@@ -189,14 +174,14 @@ test(
     accountResource = resources.find((r) => r.type === '0x1::Coin::CoinStore<0x1::TestCoin::TestCoin>');
     expect((accountResource.data as any).coin.value).toBe('0');
 
-    const token = new TypeTagStruct(StructTag.fromString('0x1::TestCoin::TestCoin'));
+    const token = new TxnBuilderTypes.TypeTagStruct(TxnBuilderTypes.StructTag.fromString('0x1::TestCoin::TestCoin'));
 
-    const scriptFunctionPayload = new TransactionPayloadScriptFunction(
-      ScriptFunction.natual(
+    const scriptFunctionPayload = new TxnBuilderTypes.TransactionPayloadScriptFunction(
+      TxnBuilderTypes.ScriptFunction.natual(
         '0x1::Coin',
         'transfer',
         [token],
-        [bcsToBytes(AccountAddress.fromHex(account4.address())), bcsSerializeUint64(123)],
+        [BCS.bcsToBytes(TxnBuilderTypes.AccountAddress.fromHex(account4.address())), BCS.bcsSerializeUint64(123)],
       ),
     );
 
@@ -205,23 +190,26 @@ test(
       client.getChainId(),
     ]);
 
-    const rawTxn = new RawTransaction(
-      AccountAddress.fromHex(mutisigAccountAddress),
+    const rawTxn = new TxnBuilderTypes.RawTransaction(
+      TxnBuilderTypes.AccountAddress.fromHex(mutisigAccountAddress),
       BigInt(sequnceNumber),
       scriptFunctionPayload,
       1000n,
       1n,
       BigInt(Math.floor(Date.now() / 1000) + 10),
-      new ChainId(chainId),
+      new TxnBuilderTypes.ChainId(chainId),
     );
 
-    const txnBuilder = new TransactionBuilderMultiEd25519((signingMessage: SigningMessage) => {
+    const txnBuilder = new TransactionBuilderMultiEd25519((signingMessage: TxnBuilderTypes.SigningMessage) => {
       const sigHexStr1 = account1.signBuffer(signingMessage);
       const sigHexStr3 = account3.signBuffer(signingMessage);
-      const bitmap = MultiEd25519Signature.createBitmap([0, 2]);
+      const bitmap = TxnBuilderTypes.MultiEd25519Signature.createBitmap([0, 2]);
 
-      const muliEd25519Sig = new MultiEd25519Signature(
-        [new Ed25519Signature(sigHexStr1.toUint8Array()), new Ed25519Signature(sigHexStr3.toUint8Array())],
+      const muliEd25519Sig = new TxnBuilderTypes.MultiEd25519Signature(
+        [
+          new TxnBuilderTypes.Ed25519Signature(sigHexStr1.toUint8Array()),
+          new TxnBuilderTypes.Ed25519Signature(sigHexStr3.toUint8Array()),
+        ],
         bitmap,
       );
