+++ ecosystem/typescript/sdk/src/transaction_builder/aptos_types/index.ts
@@ -1,3 +1,5 @@
+import { Buffer } from 'buffer/';
+
 export * from './account_address';
 export * from './authenticator';
 export * from './transaction';
@@ -5,3 +7,6 @@ export * from './type_tag';
 export * from './identifier';
 export * from './ed25519';
 export * from './multi_ed25519';
+export * from './authentication_key';
+
+export type SigningMessage = Buffer;
